+++ src/main/steps/c100-rebuild/other-proceedings/documentUpload/getController.test.ts
@@ -5,6 +5,8 @@ import { mockResponse } from '../../../../../test/unit/utils/mockResponse';
 import { FieldPrefix } from '../../../../app/case/case';
 
 import EmergencyDocumentUpload from './getController';
+import { C100_OTHER_PROCEEDINGS_DOCUMENT_UPLOAD } from '../../../urls';
+import { applyParms } from '../../../common/url-parser';
 
 jest.mock('axios');
 const mockedAxios = axios as jest.Mocked<typeof axios>;
@@ -42,6 +44,8 @@ describe('DocumentUpload Get Controller', () => {
       orderType: 'otherOrder',
       orderId: '1',
     };
+    req.originalUrl = applyParms(C100_OTHER_PROCEEDINGS_DOCUMENT_UPLOAD, {orderType: 'otherOrder',
+      orderId: '1'});
 
     await controller.get(req, res);
 
@@ -148,6 +152,8 @@ describe('DocumentUpload Get Controller', () => {
         orderType: 'otherOrder',
         orderId: '1',
       };
+      req.originalUrl = applyParms(C100_OTHER_PROCEEDINGS_DOCUMENT_UPLOAD, {orderType: 'otherOrder',
+      orderId: '1'});
 
       await controller.get(req, res);
